+++ resources/js/modules/autosave/mixins.js
@@ -1,0 +1,6 @@
+import CloseMixin from "./closeMixin";
+import AutosaveMixin from "./autosaveMixin";
+
+const mixins = [CloseMixin, AutosaveMixin];
+
+export default mixins;
