+++ resources/js/processes/scripts/components/ScriptEditor.vue
@@ -188,13 +188,14 @@ import _ from "lodash";
 import TopMenu from "../../../components/Menu.vue";
 // eslint-disable-next-line no-unused-vars
 import customFilters from "../customFilters";
-import { showLeaveWarning } from "../../../leave-warning";
+import autosaveMixins from "../../../modules/autosave/mixins";
 
 export default {
   components: {
     MonacoEditor,
     TopMenu,
   },
+  mixins: [...autosaveMixins],
   props: {
     script: {
       type: Object,
@@ -269,11 +270,33 @@ export default {
           },
         ],
       },
+      closeHref: "/designer/scripts",
     };
   },
   computed: {
     language() {
       return this.scriptExecutor.language;
+    },
+    autosaveApiCall() {
+      return () => {
+        this.setLoadingState(true);
+        ProcessMaker.apiClient
+          .put(`scripts/${this.script.id}/draft`, {
+            code: this.code,
+            title: this.script.title,
+            description: this.script.description,
+            script_executor_id: this.script.script_executor_id,
+            run_as_user_id: this.script.run_as_user_id,
+            timeout: this.script.timeout,
+          })
+          .then(() => {
+            this.setVersionIndicator(true);
+            window.ProcessMaker.EventBus.$emit("save-changes");
+          })
+          .finally(() => {
+            this.setLoadingState(false);
+          });
+      };
     },
   },
   watch: {
@@ -400,53 +423,6 @@ export default {
             onError(err);
           }
         });
-    },
-    onClose() {
-      window.removeEventListener("beforeunload", showLeaveWarning);
-      const forceAutosave = true;
-      this.handleAutosave(forceAutosave).then(() => {
-        window.location.href = "/designer/scripts";
-      }).catch(() => {
-        window.addEventListener("beforeunload", showLeaveWarning);
-      });
-    },
-    async handleAutosave(force = false) {
-      if (this.isVersionsInstalled === false) {
-        return;
-      }
-
-      const apiCall = () => {
-        this.setLoadingState(true);
-        ProcessMaker.apiClient
-          .put(`scripts/${this.script.id}/draft`, {
-            code: this.code,
-            title: this.script.title,
-            description: this.script.description,
-            script_executor_id: this.script.script_executor_id,
-            run_as_user_id: this.script.run_as_user_id,
-            timeout: this.script.timeout,
-          })
-          .then(() => {
-            // Set draft status.
-            this.setVersionIndicator(true);
-            window.ProcessMaker.EventBus.$emit("save-changes");
-          })
-          .finally(() => {
-            this.setLoadingState(false);
-          });
-      };
-
-      if (force) {
-        apiCall();
-      } else {
-        if (this.debounceTimeout) {
-          clearTimeout(this.debounceTimeout);
-        }
-
-        this.debounceTimeout = setTimeout(() => {
-          apiCall();
-        }, this.autoSaveDelay);
-      }
     },
     discardDraft() {
       ProcessMaker.apiClient
